[#78] test: count 테스트 코드 작성

+++ src/main/java/com/sparta/outsourcing/domain/store/repository/StoreRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
-    Long countByOwnerAndDeletedIsFalse(User owner);
+    int countByOwnerAndDeletedIsFalse(User owner);
     Optional<Store> findByName(String name);
     Page<Store> findAllByDeletedIsFalse(Pageable pageable);
     Page<Store> findAllByDeletedIsFalseAndNameContaining(String name, Pageable pageable);
